refactor: move jest transformer class to package entry
BREAKING CHANGE:
- One currently refers type in `jest.config.js`
```
/** @typedef {import('ts-jest')} */
module.exports = {
  //...
}
```

should change to
```
/** @typedef {import('ts-jest/dist/types')} */
module.exports = {
  //...
}
```

- Remove possibilities to import `mocked`, `createJestPreset`, `pathsToModuleNameMapper` from package entry. One should change to
```
import { mocked, createJestPreset, pathsToModuleNameMapper` } from 'ts-jest/utils'
```
